+++ isitfit/__init__.py
@@ -1,1 +1,1 @@
-isitfit_version='0.8.4'
+isitfit_version='0.9.0'